requirements.txt
gradio==3.50.2
numpy==1.25.1
Pillow==10.0.0
scipy==1.11.1
streamlit==1.28.0
streamlit-drawable-canvas==0.9.3
